requirements.txt
streamlit
beautifulsoup4
requests
chromadb
sentence-transformers
google-generativeai
python-dotenv
